+++ tests/www/api/experimental/test_endpoints.py
@@ -56,6 +56,10 @@ class TestApiExperimental(TestBase):
         session.commit()
         session.close()
 
+        dagbag = DagBag(include_examples=True)
+        for dag in dagbag.dags.values():
+            dag.sync_to_db()
+
     def setUp(self):
         super().setUp()
